merged_project (1)/requirements.txt
Flask
Flask-Mail
Flask-SQLAlchemy
Werkzeug
python-docx
PyPDF2
scikit-learn
numpy
scipy
gunicorn